# runtime-acm/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.1
info:
  title: Api Documentation
  description: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: https://{server}
    variables:
      server:
        default: onap/acm/v3
        description: This value is assigned by the service provider
...
paths:
...
  /compositions:
    get:
      tags:
      - Automation Composition Definition
      summary: Query the commissioned automation composition definitions
        definitions
      description: Query the commissioned automation composition
        definitions, returning the automation composition details
      operationId: queryCompositionDefinitions
      parameters:
      - name: name
        in: query
        required: false
        description: Automation composition definition name. Regular expressions are supported for filtering. If
          this parameter is not specified, all automation composition definitions are returned.
        schema:
          type: string
      - name: version
        in: query
        required: false
        description: Automation composition definition version. Regular expressions are supported for filtering. If this
          parameter is not specified, all automation composition definitions that match the "name" filter are are returned.
        schema:
          type: string
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [ToscaServiceTemplates](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplates.java)
            that contains the automation composition definitions that match the requested filters.
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplates'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getAllCompositionDefinitions.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplates'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getAllCompositionDefinitions.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    post:
      tags:
      - Automation Composition Definition
      summary: Commissions automation composition definitions
      description: Commissions automation composition definitions, returning the UUIDs of
        automation composition definitions commissioned by this request.
      operationId: createCompositionDefinitions
      parameters:
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
            containing the automation composition definitions to be commissioned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionDefinitions.json'
          application/yaml:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCompositionDefinitions.yaml'
        required: true
      responses:
        201:
          description: Serialised instance of
            [CommissioningResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/commissioning/CommissioningResponse.java)
            containing the UUIDs of automation composition definitions created by this request
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCommissionCompositionDefinitionsResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/postCommissionCompositionDefinitionsResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
  /compositions/{compositionId}:
    get:
      tags:
      - Automation Composition Definition
      summary: Get details of the requested commissioned automation composition
        definitions
      description: Get details of the requested commissioned automation composition
        definitions, returning all automation composition details
      operationId: getCompositionDefinition
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition to get
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description:
            Serialised instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
            containing the requested automation composition definition.
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getSingleCompositionDefinition.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/getSingleCompositionDefinition.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: Specified automation composition definition not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
    put:
      tags:
      - Automation Composition Definition
      summary: Primes or deprimes an automation composition definition
      description: Primes or deprimes an automation composition definition by sending the AC Element Types to participants and
        getting participants to take responsibility for AC Element Types in this AC Type.
      operationId: compositionDefinitionPriming
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Serialised instance of
            [AcTypeStateUpdate](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/commissioning/AcTypeStateUpdate.java)
            which specifies the requested state change on the automation concept instance
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcTypeStateUpdate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/putAcTypeStateUpdate.json'
          application/yaml:
            schema:
              $ref: '#/components/schemas/AcTypeStateUpdate'
            example:
              externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/putAcTypeStateUpdate.yaml'
      responses:
        202:
          description: Accepted, the request has been accepted and forwarded to participants
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: Specified automation composition definition not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
      x-codegen-request-body-name: body
    delete:
      tags:
      - Automation Composition Definition
      summary: Delete a commissioned automation composition definition
      description: Deletes a commissioned automation composition definition,returning the UUID of the automation composition definition deleted by this request
      operationId: deleteCompositionDefinition
      parameters:
      - name : compositionId
        in: path
        description: The UUID of the automation composition definition to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-onap-RequestId
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: Serialised instance of
            [CommissioningResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/commissioning/CommissioningResponse.java)
            containing the UUID of the automation composition deleted by this request
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/deleteCompositionDefinitionResponse.json'
            application/yaml:
              schema:
                $ref: '#/components/schemas/CommissioningResponse'
              example:
                externalValue: 'https://raw.githubusercontent.com/onap/policy-clamp/master/runtime-acm/src/main/resources/openapi/examples/deleteCompositionDefinitionResponse.yaml'
        401:
          description: Authentication Error, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        404:
          description: Specified automation composition definition not found, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
        400:
          description: Bad Request, returns an instance of
            [SimpleResponse](https://github.com/onap/policy-clamp/blob/master/models/src/main/java/org/onap/policy/clamp/models/acm/messages/rest/SimpleResponse.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleResponse'
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: London
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    ToscaServiceTemplates:
      title: ToscaServiceTemplates
      type: object
    ToscaServiceTemplate:
      title: ToscaServiceTemplate
      type: object
...
    SimpleResponse:
      title: SimpleResponse
      type: object
    CommissioningResponse:
      title: CommissioningResponse
      type: object
    AcTypeStateUpdate:
      title: AcTypeStateUpdate
      type: object
...